test(analytics): rename PageViewTracker multi-provider suite to match repo convention
